ajustes pontuais em alguns itens do backend do admin e front da aplic…

+++ server/api/admin/medias/index.delete.js
@@ -24,39 +24,41 @@ export default defineEventHandler(async (event) => {
 
 		if (Boolean(media)) {
 			if (media.type === config.typesMedia[3]) {
-				media.value.forEach(async (mediaFile) => {
-					// const hasMediaExists = fs.existsSync(`public/uploads/${mediaFile}`);
+				if (media.value) {
+					media.value.forEach(async (mediaFile) => {
+						// const hasMediaExists = fs.existsSync(`public/uploads/${mediaFile}`);
 
-					// if (hasMediaExists) fs.unlinkSync(`public/uploads/${mediaFile}`);
+						// if (hasMediaExists) fs.unlinkSync(`public/uploads/${mediaFile}`);
 
-					const bucket = googleCloudStorage.bucket(config.gcsBucketname);
-					const fileUp = bucket.file(
-						`${config.gcsSubfolder}${config.gcsSubfolderEnvironment}${mediaFile}`
-					);
+						const bucket = googleCloudStorage.bucket(config.gcsBucketname);
+						const fileUp = bucket.file(
+							`${config.gcsSubfolder}${config.gcsSubfolderEnvironment}${mediaFile}`
+						);
 
-					try {
-						await fileUp.delete();
-						console.log('Arquivo excluído com sucesso');
-					} catch (error) {
-						console.error(`Erro ao excluir a imagem: ${error}`);
-					}
+						try {
+							await fileUp.delete();
+							console.log('Arquivo excluído com sucesso');
+						} catch (error) {
+							console.error(`Erro ao excluir a imagem: ${error}`);
+						}
 
-					// const stream = fileUp.createWriteStream({
-					// 	metadata: {
-					// 		contentType: file.mimetype, // Tipo MIME do arquivo
-					// 	},
-					// });
+						// const stream = fileUp.createWriteStream({
+						// 	metadata: {
+						// 		contentType: file.mimetype, // Tipo MIME do arquivo
+						// 	},
+						// });
 
-					// stream.end(fs.readFileSync(file.filepath));
+						// stream.end(fs.readFileSync(file.filepath));
 
-					// stream.on('error', (error) => {
-					// 	console.error(`Erro ao enviar arquivo para o GCS: ${error}`);
-					// });
+						// stream.on('error', (error) => {
+						// 	console.error(`Erro ao enviar arquivo para o GCS: ${error}`);
+						// });
 
-					// stream.on('finish', () => {
-					// 	console.log('Arquivo enviado com sucesso para o GCS');
-					// });
-				});
+						// stream.on('finish', () => {
+						// 	console.log('Arquivo enviado com sucesso para o GCS');
+						// });
+					});
+				}
 			}
 
 			await media.destroy();
